requirements.txt
boto3
pyyaml
flake8
pytest